删除对 url 的重复判断
