+++ registry/lib/components/feeds/pgc-watchlater/index.ts
@@ -1,6 +1,34 @@
 import { defineComponentMetadata } from '@/components/define'
 import { feedsUrls } from '@/core/utils/urls'
-import { entry } from './entry'
+import { forEachFeedsCard } from '@/components/feeds/api'
+import { urlChange } from '@/core/observer'
+import { getUID } from '@/core/utils'
+import WatchlaterButton from './WatchlaterButton.vue'
+
+const feedPgcCardSelector = 'a.bili-dyn-card-pgc'
+const feedPgcCardCoverSelector = '.bili-dyn-card-pgc__cover'
+
+const entry = async () => {
+  urlChange(() => {
+    if (!getUID()) {
+      return
+    }
+    forEachFeedsCard({
+      added: card => {
+        const videoCard: HTMLAnchorElement = card.element.querySelector(feedPgcCardSelector)
+        if (videoCard) {
+          if (videoCard.querySelector('.bili-dyn-card-pgc__mark')) {
+            return
+          }
+          const vm = new (Vue.extend(WatchlaterButton))()
+          vm.$mount()
+          const button = vm.$el
+          videoCard.querySelector(feedPgcCardCoverSelector)?.appendChild(button)
+        }
+      },
+    })
+  })
+}
 
 const author = {
   name: 'WhiteTeal55',
